# Componente A proyecto final/demo/src/main/java/com/umg/edu/gt/dev/demo/controller/PedidoController.java
package com.umg.edu.gt.dev.demo.controller;

import java.util.Optional;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.context.request.NativeWebRequest;

import com.umg.edu.gt.dev.model.CalcularTotalPedido200Response;
import com.umg.edu.gt.dev.model.ListarPedidos200Response;
import com.umg.edu.gt.dev.model.Pedido;
import com.umg.edu.gt.dev.model.PedidoInput;
import com.umg.edu.gt.dev.spec.PedidosApi;

import jakarta.validation.Valid;

/**
 * Controlador REST para gestión de Pedidos
 * Implementa la interfaz generada por OpenAPI
 */
@RestController
public class PedidoController implements PedidosApi {

    @Override
    public Optional<NativeWebRequest> getRequest() {
        return PedidosApi.super.getRequest();
    }

    /**
     * POST /api/v1/pedidos - Crear un pedido
     */
    @Override
    public ResponseEntity<Pedido> crearPedido(@Valid PedidoInput pedidoInput) {
        // TODO: Implementar lógica de creación
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

    /**
     * GET /api/v1/pedidos - Listar todos los pedidos
     */
    @Override
    public ResponseEntity<ListarPedidos200Response> listarPedidos(
            Long clienteId,
            String estado,
            Integer page,
            Integer size) {
        // TODO: Implementar lógica de listado
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

    /**
     * GET /api/v1/pedidos/{id} - Obtener un pedido por ID
     */
    @Override
    public ResponseEntity<Pedido> obtenerPedidoPorId(Long id) {
        // TODO: Implementar lógica de búsqueda por ID
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

    /**
     * PUT /api/v1/pedidos/{id} - Actualizar un pedido
     */
    @Override
    public ResponseEntity<Pedido> actualizarPedido(
            Long id,
            @Valid PedidoInput pedidoInput) {
        // TODO: Implementar lógica de actualización
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

    /**
     * GET /api/v1/pedidos/{id}/calcular-total - Calcular total del pedido
     */
    @Override
    public ResponseEntity<CalcularTotalPedido200Response> calcularTotalPedido(Long id) {
        // TODO: Implementar cálculo usando Componente C
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }
}