# Componente A proyecto final/demo/src/main/resources/ClienteApiSpecification.yaml
openapi: 3.0.1
info:
  title: API Componente A - Clientes y Pedidos
  version: 1.0.0
  description: API para gestion de clientes y pedidos (MariaDB)
servers:
  - url: http://localhost:8080
    description: Servidor local Componente A

paths:
  /api/v1/clientes:
    post:
      tags:
        - Clientes
      summary: Crear un cliente
      operationId: crearCliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '201':
          description: Cliente creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '400':
          description: Datos invalidos
    get:
      tags:
        - Clientes
      summary: Listar todos los clientes
      operationId: listarClientes
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
                  
  /api/v1/clientes/{id}:
    get:
      tags:
        - Clientes
      summary: Obtener un cliente por ID
      operationId: obtenerClientePorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente no encontrado

  /api/v1/pedidos:
    post:
      tags:
        - Pedidos
      summary: Crear un pedido
      operationId: crearPedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoInput'
      responses:
        '201':
          description: Pedido creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          description: Datos invalidos
    get:
      tags:
        - Pedidos
      summary: Listar todos los pedidos
      operationId: listarPedidos
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'
                  
  /api/v1/pedidos/{id}:
    get:
      tags:
        - Pedidos
      summary: Obtener un pedido por ID
      operationId: obtenerPedidoPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '404':
          description: Pedido no encontrado

components:
  schemas:
    Cliente:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        codigoUnico:
          type: string
          
    ClienteInput:
      type: object
      required:
        - nombre
        - email
      properties:
        nombre:
          type: string
          minLength: 3
          maxLength: 100
        email:
          type: string
          format: email
        telefono:
          type: string
        direccion:
          type: string
          
    ItemPedido:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productoNombre:
          type: string
        cantidad:
          type: integer
          minimum: 1
        precioUnitario:
          type: number
          format: double
        subtotal:
          type: number
          format: double
          
    ItemPedidoInput:
      type: object
      required:
        - productoNombre
        - cantidad
        - precioUnitario
      properties:
        productoNombre:
          type: string
        cantidad:
          type: integer
          minimum: 1
        precioUnitario:
          type: number
          format: double
          minimum: 0.01
          
    Pedido:
      type: object
      properties:
        id:
          type: integer
          format: int64
        clienteId:
          type: integer
          format: int64
        codigoPedido:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedido'
        subtotal:
          type: number
          format: double
        impuestos:
          type: number
          format: double
        total:
          type: number
          format: double
          
    PedidoInput:
      type: object
      required:
        - clienteId
        - items
      properties:
        clienteId:
          type: integer
          format: int64
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ItemPedidoInput'